// src/shared/types/courses.ts
export type Course = {
  id: number;
  courseId: string;
  title: string;
  description: string;
  detail: string;
  requirements: string;
  materials: string;
  price_per_hour: string;
  price_per_dollar: string;
  price_per_toman: string;
  language: string;
  level: string;
  schedule_day: string;
  schedule_start: string;
  schedule_end: string;
  capacity: number;
  active_students: number;
  length: number;
  course_duration: number;
  image: string;
  language_flag: string;
  lessons: unknown[];
  tutor: {
    id: number;
    user: number;
    profile_picture: string | null;
    languages_spoken: string[] | Record<string, string>;
    subjects: unknown[];
  };
};